Class "ChIPScore"

ChIPScore is the result of the processingChIP function. This object contains the extracted ChIP Score from ChIP data, the loci of interest and optional paramters associated to ChIPanalyser. The loci of interest will either be user provided or the top n regions as defined by the reduce argument im processingChIP. This object has the sole purpose of aiding the storage and parsing of data and parameters.

Objects from the Class

Object of this class are created internaly and will be parsed to other objects as is.

Slots

scores:

Object of class "list" List of extracted ChIP scores

loci:

Object of class "loci" GRanges containing loci of interest

ploidy:

Object of class "numeric" Ploidy level of the organism

boundMolecules:

Object of class "vector" Number of Bound molecules to the DNA

backgroundSignal:

Object of class "numeric" ChIP background signal (average ChIP score)

maxSignal:

Object of class "numeric" max ChIP signal

lociWidth:

Object of class "numeric" Width of loci if reduce is used and no loci are provided

chipMean:

Object of class "numeric" Average ChIP peak width

chipSd:

Object of class "numeric" Standard Deviation of ChIP peak width

chipSmooth:

Object of class "vector" Smoothing window width for ChIP score

stepSize:

Object of class "numeric" Defining resolution size of ChIP like profiles (10bp = signal will be only considered every 10bp)

removeBackground:

Object of class "numeric" Signal Threshold to be removed. Default removes all negative scores

noiseFilter:

Object of class "character" Type of noise filter to be used on ChIP data.

PWMThreshold:

Object of class "numeric" Threshold of PWM scores that will be selected

strandRule:

Object of class "character" Rule to compute strand score (max, mean or sum)

whichstrand:

Object of class "character" Which strand should be used to compute PWM scores.

lambdaPWM:

Object of class "vector" Lambda value - Scaling factor to the PWM

naturalLog:

Object of class "logical" PFM to PWM conversion log transform ( natural log or log2)

noOfSites:

Object of class "nos" Number of Sites in the PWM that should be used to compute PWM scores.

PWMpseudocount:

Object of class "numeric" PWM pseudocount value for PFM to PWM conversion.

paramTag:

Object of class "character" Internal Tag - Code progression

Extends

Class "parameterOptions", directly.

Methods

.loci<-

signature(object = "ChIPScore", value = "loci"): ...

.scores<-

signature(object = "ChIPScore", value = "list"): ...

initialize

signature(.Object = "ChIPScore"): ...

loci

signature(object = "ChIPScore"): ...

scores

signature(object = "ChIPScore"): ...

show

signature(object = "ChIPScore"): ...

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94. Patrick C.N. Martin and Nicolae Radu Zabe (2020) Dissecting the binding mechanisms of transcription factors to DNA using a statistical thermodynamics framework. CSBJ, 18, 3590-3605.

Author

Patrick C.N. Martin

See also

Examples

showClass("ChIPScore")
#> Class "ChIPScore" [package "ChIPanalyser"]
#> 
#> Slots:
#>                                                                           
#> Name:            scores             loci           ploidy   boundMolecules
#> Class:             list             loci          numeric           vector
#>                                                                           
#> Name:  backgroundSignal        maxSignal        lociWidth         chipMean
#> Class:          numeric          numeric          numeric          numeric
#>                                                                           
#> Name:            chipSd       chipSmooth         stepSize removeBackground
#> Class:          numeric           vector          numeric          numeric
#>                                                                           
#> Name:       noiseFilter     PWMThreshold       strandRule      whichstrand
#> Class:        character          numeric        character        character
#>                                                                           
#> Name:         lambdaPWM       naturalLog        noOfSites   PWMpseudocount
#> Class:           vector          logical              nos          numeric
#>                        
#> Name:          paramTag
#> Class:        character
#> 
#> Extends: "parameterOptions"