Class "GRList"

Virutal Class to handle multiple data types for one slot ( profiles)

Objects from the Class

A virtual Class: No objects may be created from it.

Methods

GRList-class

The purpose of this virtual classe is to store data of two different formats in one slot: GRangesList and Lists

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94.

Author

Patrick C. N. Martin <p.martin@essex.ac.uk>

Examples

showClass("GRList")
#> Virtual Class "GRList" [package "ChIPanalyser"]
#> 
#> No Slots, prototype of class "CompressedGRangesList"
#> 
#> Known Subclasses: 
#> Class "CompressedGRangesList", directly
#> Class "list", directly
#> Class "GRanges", directly
#> Class "data.frame", by class "list", distance 2
#> Class "namedList", by class "list", distance 2
#> Class "GPos", by class "GRanges", distance 2
#> Class "listOfMethods", by class "list", distance 3
#> Class "UnstitchedGPos", by class "GRanges", distance 3
#> Class "StitchedGPos", by class "GRanges", distance 3