Computing ChIP-seq like profiles from Occupancy data.

computeChIPProfile compute ChIP-seq like profile from occupancy data. Occupancy data is computed using computeOccupancy.

Usage

computeChIPProfile(genomicProfiles, loci, parameterOptions = NULL,
    norm = TRUE, method = c("moving_kernel","truncated_kernel","exact"),
    peakSignificantThreshold= NULL,cores=1, verbose = TRUE)

Arguments

genomicProfiles

genomicProfiles is the result of computeOccupancy. This object should be a genomicProfiles object.

loci

loci is either a GRanges or ChIPScore object. ChIPScore-class will be the result of processingChIP. This object represents the set of Loci you are interested in analysing. If you have followed the full ChIPanalyser pipe line, you would have used the processingChIP function that would return a ChIPScore-class object containing your loci of interest. GRanges are also supported if you are only using part of the pipeline.

parameterOptions

parameterOptions is a parameterOptions object. This object is used to store the numerous paramters offered by ChIPanalyser. This argument is optional as all arguments are also parse in both ChIPScore-class and genomicProfiles objects. If you wanted to make some last minute changes, parameterOptions is the way to go. We recomend that you set your desired options before hand.

norm

norm is a logical value. If TRUE, the ChIP-seq like profile will be normalised towards maximum Occupancy. If FALSE, the profile will be left as is.

method

method is a character string of one of the following: c("moving_kernel","truncated_kernel","exact"). If set to moving_kernel, the peaks will be approximated using Rcpp (Default). If set to truncated_kernel, the peaks will be approximated however this method does not require Rcpp. If set to exact, the peaks will not be approximated.

peakSignificantThreshold

peakSignificantThreshold is a threshold at which peaks will be selected. IMPORTANT: if you select "moving_kernel" as described in method then this threshold is a numeric value describing the peak tail hight cutoff value (Default = 0.001). In the case of "truncated_kernel" and "exact", the threhsold represents a distance in base pair from the peak summit at which the peak should be cut (Default = 1250). The default is set to NULL in this function. This just means that either the value is provided bu user with the appropriate method. If not, the default will be selected depending on the method selected.

cores

cores is the number of cores that will be used to compute ChIP profiles.

verbose

verbose is a logical value. If TRUE, progress messages will be displayed in console. If FALSE, no progress messages will be dispalyed in console.

Details

computeChIPProfile converts Transcription Factor occuapncy to a profile resembling the one of a ChIP-seq profile. Internally a few paramters are required to build a ChIP like profile. These parameters are either defined and stored in a ChIPScore object (Paramters are updated based on your ChIP data ), a genomicProfiles (user defined at the start of the analysis) or a parameterOptions (if you want to update values as you go along)

Value

Returns a genomicProfiles objec containing all ChIP-seq like profile for every combination of lambdaPWM and boundMolecules provided by the user.

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94. Patrick C.N. Martin and Nicolae Radu Zabe (2020) Dissecting the binding mechanisms of transcription factors to DNA using a statistical thermodynamics framework. CSBJ, 18, 3590-3605.

Author

Patrick C.N. Martin <pcnmartin@gmail.com>

Examples

#Extracting Data
data(ChIPanalyserData)
# path to Position Frequency Matrix
PFM <- file.path(system.file("extdata",package="ChIPanalyser"),"BEAF-32.pfm")
#As an example of genome, this example will run on the Drosophila genome
# \donttest{
if(!require("BSgenome.Dmelanogaster.UCSC.dm6", character.only = TRUE)){
    if (!requireNamespace("BiocManager", quietly=TRUE))
        install.packages("BiocManager")
    BiocManager::install("BSgenome.Dmelanogaster.UCSC.dm6")
    }
library(BSgenome.Dmelanogaster.UCSC.dm6)
DNASequenceSet <- getSeq(BSgenome.Dmelanogaster.UCSC.dm6)
# Building genomicProfiles object
GPP <- genomicProfiles(PFM=PFM, PFMFormat="JASPAR",BPFrequency=DNASequenceSet)

# Computing Genome Wide
GenomeWide <- computeGenomeWideScores(genomicProfiles = GPP,
    DNASequenceSet = DNASequenceSet)
#> Extracting genome wide scores 
#> Whole Genome ~ Both strands 
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Computing Mean waiting time 

#Compute PWM Scores
PWMScores <- computePWMScore(genomicProfiles = GenomeWide,
     DNASequenceSet = DNASequenceSet, loci = top, chromatinState = Access)
#> PWM Scores Extraction
#Compute Occupnacy
Occupancy <- computeOccupancy(genomicProfiles = PWMScores)
#> Computing Occupancy at sites higher than threshold. 

#Compute ChIP profiles
chipProfile <- computeChIPProfile(genomicProfiles=Occupancy,loci=top)
#> Computing ChIP Profile 
chipProfile
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>                         genomicProfiles object: ChIP like Profiles 
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>  Parameter combination:  lambda = 1 & boundMolecules = 1000 
#>  
#> $`lambda = 1 & boundMolecules = 1000`
#> GRangesList object of length 4:
#> $`chr3L:20300001..20320000`
#> GRanges object with 2000 ranges and 1 metadata column:
#>                            seqnames            ranges strand |        ChIP
#>                               <Rle>         <IRanges>  <Rle> |   <numeric>
#>   chr3L:20300001..20320000    chr3L 20300001-20300010      * | 1.57008e-05
#>   chr3L:20300001..20320000    chr3L 20300011-20300020      * | 1.65058e-05
#>   chr3L:20300001..20320000    chr3L 20300021-20300030      * | 1.73521e-05
#>   chr3L:20300001..20320000    chr3L 20300031-20300040      * | 1.68246e-05
#>   chr3L:20300001..20320000    chr3L 20300041-20300050      * | 1.61817e-05
#>                        ...      ...               ...    ... .         ...
#>   chr3L:20300001..20320000    chr3L 20319951-20319960      * | 2.50940e-10
#>   chr3L:20300001..20320000    chr3L 20319961-20319970      * | 2.38702e-10
#>   chr3L:20300001..20320000    chr3L 20319971-20319980      * | 2.27060e-10
#>   chr3L:20300001..20320000    chr3L 20319981-20319990      * | 2.15986e-10
#>   chr3L:20300001..20320000    chr3L 20319991-20320000      * | 2.05453e-10
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
#> 
#> ...
#> <3 more elements>
#> 
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>                         Associated parameter Options 
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>  lambdaPWM:                                                 1
#>  boundMolecules:                                         1000
#>  stepSize:                                                 10
#>  backgroundSignal:                                          0
#>  maxSignal:                                                 1
#>  removeBackground:                                          0
#>  chipMean:                                                200
#>  chipSd:                                                  200
#>  chipSmooth:                                              250
# }