Compute PWM Scores of sites above threshold.

computePWMScore will compute and extract all sites that exhibit a PWM Score higher than a threshold. This threshold (see PWMThreshold) will determine the percentage of total sites that should NOT be considered.

Usage

computePWMScore(genomicProfiles,DNASequenceSet,
    loci = NULL, chromatinState = NULL,parameterOptions=NULL,cores=1, verbose = TRUE)

Arguments

DNASequenceSet

DNASequenceSet is a DNAStringSet or a BSgenome containing the full sequence of the organism of interest.

genomicProfiles

genomicProfiles is a genomicProfiles object resulting from the computeGenomeWideScores function.

loci

loci is a GRanges object containing the Loci of interest or a ChIPScore object result of processingChIP function.

parameterOptions

parameterOptions is a parameterOptions object containing parameters that you wish to parse/change when computing PWMScores.

chromatinState

chromatinState is a GRanges object sites of accessible DNA or DNA affinity scores.

cores

cores is the number of cores used to compute PWM Scores.

verbose

verbose is a logical value indicating if progress messages should be printed or not.

Details

After determining genome wide scores, it is possible to only compute and extract high affinity sites (in the sense that they have a high PWM Score). If a PWMThreshold is not set by user, the default value is set at 0.7. This means that 70 % of sites will NOT be selected. Only the top 30 % will be computed and extracted. If one is interested in all PWM Scores at a genome wide scale ( or accessible DNA ), this is possible by setting PWMThreshold to zero.

Value

computePWMScore will return a genomicProfiles object. The profiles slot will have been updated. This slot will now contain a GRangesList

with each element being a GRanges. This GRanges will contain postion of each sites (start, end and strand) and the PWMScore associated to that site.

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94.

Author

Patrick C.N. Martin <pcnmartin@gmail.com>

Examples

#Data extraction
data(ChIPanalyserData)
# path to Position Frequency Matrix
PFM <- file.path(system.file("extdata",package="ChIPanalyser"),"BEAF-32.pfm")
#As an example of genome, this example will run on the Drosophila genome
# \donttest{
if(!require("BSgenome.Dmelanogaster.UCSC.dm6", character.only = TRUE)){
    if (!requireNamespace("BiocManager", quietly=TRUE))
        install.packages("BiocManager")
    BiocManager::install("BSgenome.Dmelanogaster.UCSC.dm6")
    }
library(BSgenome.Dmelanogaster.UCSC.dm6)
DNASequenceSet <- getSeq(BSgenome.Dmelanogaster.UCSC.dm6)
chip<-processingChIP(chip,top)
#Building data objects
GPP <- genomicProfiles(PFM=PFM,PFMFormat="JASPAR",BPFrequency=DNASequenceSet)

# Computing Genome Wide
GenomeWide <- computeGenomeWideScores(DNASequenceSet = DNASequenceSet,
    genomicProfiles = GPP)
#> Extracting genome wide scores 
#> Whole Genome ~ Both strands 
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Computing Mean waiting time 

#Compute PWM Scores
PWMScores <- computePWMScore(DNASequenceSet = DNASequenceSet,
    genomicProfiles = GenomeWide,
    loci  = chip, chromatinState = Access)
#> PWM Scores Extraction
PWMScores
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>                         genomicProfiles object: PWM scores above Threshold 
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>  Scores above Threshold in locus:  
#> GRanges object with 162 ranges and 2 metadata columns:
#>                            seqnames            ranges strand |  PWMScore
#>                               <Rle>         <IRanges>  <Rle> | <numeric>
#>   chr3L:20300001..20320000    chr3L 20300005-20300019      - | -0.260743
#>   chr3L:20300001..20320000    chr3L 20300015-20300029      + | -3.581791
#>   chr3L:20300001..20320000    chr3L 20300022-20300036      - |  2.266402
#>   chr3L:20300001..20320000    chr3L 20300066-20300080      + | -7.941449
#>   chr3L:20300001..20320000    chr3L 20300293-20300307      + | -6.443234
#>                        ...      ...               ...    ... .       ...
#>   chr3L:20300001..20320000    chr3L 20318562-20318576      - |  -3.48805
#>   chr3L:20300001..20320000    chr3L 20318607-20318621      - |  -9.10951
#>   chr3L:20300001..20320000    chr3L 20318620-20318634      + |  -3.39948
#>   chr3L:20300001..20320000    chr3L 20318724-20318738      + |  -7.80125
#>   chr3L:20300001..20320000    chr3L 20318728-20318742      + |  -6.63695
#>                            DNAaffinity
#>                              <numeric>
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>                        ...         ...
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>   chr3L:20300001..20320000           1
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#>  3  remaining loci 
#>  Locus:                chr2L:6480001..6500000 (length = 142)
#>  Locus:                chr3R:16140001..16160000 (length = 116)
#>  Locus:                chr2L:19500001..19520000 (length = 78)
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>                         Associated parameters 
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  
#>  minPWMScore:                               -62.9823294882894
#>  maxPWMScore:                                12.9532005236194
#>  lambdaPWM:                                                 1
#>  PWMThreshold:                                            0.7
# }