Class "genomicProfiles"

genomicProfiles is an S4 object serving two purposes: (i) storing internal computed data and (ii) storing paramter options. This object is parsed through the different steps of the pipeline to facilitate that parsing and changing of paramters.

Objects from the Class

Objects can be created by calls of the form genomicProfiles(ploidy, boundMolecules, backgroundSignal, maxSignal, lociWidth, chipMean, chipSd, chipSmooth, stepSize, noiseFilter, removeBackground, lambdaPWM, PWMpseudocount, naturalLog, noOfSites, PWMThreshold, strandRule, whichstrand, PFM, PWM, PFMFormat, BPFrequency, minPWMScore, maxPWMScore, profiles, DNASequenceLength, averageExpPWMScore).

Slots

PWM:

Object of class "matrix": A Position Weight Matrix (either supplied or internally computed if PFM is provided)

PFM:

Object of class "matrix": A Position Frequency Matrix (may also be a path to file containing PFM)

PFMFormat:

Object of class "character": A character string of one of the following: raw, transfac,JASPAR or sequences

BPFrequency:

Object of class "vector": Base Pair Frequency in the genome (if a DNA sequence is provided (as a DNAStringSet or BSgenome), will be automatically computed internally). Default:c(0.25,0.25,0.25,0.25)

minPWMScore:

Object of class "vector": Lowest PWM score accros the genome (computed and updated internally)

maxPWMScore:

Object of class "vector": Highest PWM score across the genome (computed and updated internally)

profiles:

Object of class "GRList": Containins GRanges with sites above threshold and associated metrics (PWMscore and Occupancy) - Computed Internally

DNASequenceLength:

Object of class "vector": Length of the Genome (or accesible genome) - computed internally

averageExpPWMScore:

Object of class "vector": Average exponential PWM score across the genome (or accesible genome) - computed internally

ZeroBackground:

Object of class "vector": Internal background value (computed internally)

drop:

Object of class "vector": Stores Loci that do contain accesible DNA if it were to be the case (computed and updated internally)

tags:

Object of class "character" ~Internal Tags~

ploidy:

Object of class "numeric": A numeric Value descibing the ploidy of the organism. Default: 2

boundMolecules:

Object of class "vector": A vector (or single value) containing the number of bound Molecules (bound Transcription Factors): Default: 1000

backgroundSignal:

Object of class "numeric": A numeric value descibing the ChIP-seq background Signal (average signal from real ChIP seq data). Default: 0

maxSignal:

Object of class "numeric": A numeric value describing the highest ChIP-seq signal (from real ChIP-seq data). Default: 1

lociWidth:

Object of class "numeric" ~~

chipMean:

Object of class "numeric": A numeric value describing the mean width of a ChIP- seq peak. Default:150

chipSd:

Object of class "numeric": A numeric value describing the standard deviation of ChIP-seq peaks. Default: 150

chipSmooth:

Object of class "vector": A numeric value describing the width of the window used to smooth Occupancy profiles into ChIP profiles. Default:250

stepSize:

Object of class "numeric": A numeric value describing the step Size (in base pairs) between each ChIP-seq score. Default:10 (Scored every 10 base pairs)

removeBackground:

Object of class "numeric": A numeric value describing the value at which score should be removed. Defualt:0 (If negative scores then remove)

noiseFilter:

Object of class "character" ~Describes the noiseFilter method that will be applied to ChIP data (Zero, mean, median, sigmoid)~

PWMThreshold:

Object of class "numeric": Threshold at which PWM Score should be selected (only sites above threshold will be selected - between 0 and 1)

strandRule:

Object of class "character": "mean", "max" or "sum" will dertermine how strand should be handle for computing PWM Scores. Default : "max"

whichstrand:

Object of class "character": "+","-" or "+-" on which strand should PWM Score be computed. Default: "+-"

lambdaPWM:

Object of class "vector" A vector (or single value) contaning values for lambdaPWM Default:1

naturalLog:

Object of class "logical": A logical value describing if natural Log will be used to compute the PWM (if FALSE then log2 will be used). Default: TRUE

noOfSites:

Object of class "nos" A Positive integer descibing number of sites (in base pair) should be used from the PFM to compute PWM. Default =0 (Full width of binding site will be used when set to 0)

PWMpseudocount:

Object of class "numeric": A numeric value describing a PWMpseudocount for PWM computation. Default:1

paramTag:

Object of class "character" ~Internal~

Extends

Class "genomicProfilesInternal", directly. Class "parameterOptions", directly.

Methods

initialize

signature(.Object = "genomicProfiles"): ...

show

signature(object = "genomicProfiles"): ...

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94. Patrick C.N. Martin and Nicolae Radu Zabe (2020) Dissecting the binding mechanisms of transcription factors to DNA using a statistical thermodynamics framework. CSBJ, 18, 3590-3605.

Author

Partick C. N. Martin <p.martin@essex.ac.uk>

See also

Examples

showClass("genomicProfiles")
#> Class "genomicProfiles" [package "ChIPanalyser"]
#> 
#> Slots:
#>                                                                
#> Name:                 PWM                PFM          PFMFormat
#> Class:             matrix             matrix          character
#>                                                                
#> Name:         BPFrequency        minPWMScore        maxPWMScore
#> Class:             vector             vector             vector
#>                                                                
#> Name:            profiles  DNASequenceLength averageExpPWMScore
#> Class:             GRList             vector             vector
#>                                                                
#> Name:      ZeroBackground               drop               tags
#> Class:             vector             vector          character
#>                                                                
#> Name:              ploidy     boundMolecules   backgroundSignal
#> Class:            numeric             vector            numeric
#>                                                                
#> Name:           maxSignal          lociWidth           chipMean
#> Class:            numeric            numeric            numeric
#>                                                                
#> Name:              chipSd         chipSmooth           stepSize
#> Class:            numeric             vector            numeric
#>                                                                
#> Name:    removeBackground        noiseFilter       PWMThreshold
#> Class:            numeric          character            numeric
#>                                                                
#> Name:          strandRule        whichstrand          lambdaPWM
#> Class:          character          character             vector
#>                                                                
#> Name:          naturalLog          noOfSites     PWMpseudocount
#> Class:            logical                nos            numeric
#>                          
#> Name:            paramTag
#> Class:          character
#> 
#> Extends: "genomicProfilesInternal", "parameterOptions"