parameter Options object

parameterOptions is an object used to store and parse the various parameters needed throughout this analysis pipeline.

Usage

parameterOptions(ploidy = 2, boundMolecules = 1000, backgroundSignal = 0, maxSignal = 1, lociWidth = 20000, chipMean = 200, chipSd = 200, chipSmooth = 250, stepSize = 10, removeBackground = 0, noiseFilter = "zero", naturalLog = TRUE, noOfSites = "all", PWMThreshold = 0.7, strandRule = "max", whichstrand = "+-", PWMpseudocount = 1, lambdaPWM = 1)

Arguments

ploidy

ploidy: A numeric Value descibing the ploidy of the organism. Default: 2

boundMolecules

boundMolecules: A vector (or single value) containing the number of bound Molecules (bound Transcription Factors): Default: 1000

backgroundSignal

backgroundSignal: A numeric value descibing the ChIP-seq background Signal (average signal from real ChIP seq data). Default: 0

maxSignal

maxSignal: A numeric value describing the highest ChIP-seq signal (from real ChIP-seq data). Default: 1

lociWidth

lociWidth: A numeric value describing the width of the bins used to split ChIP profiles parsed to processingChIP. Default = 20000

chipMean

chipMean: A numeric value describing the mean width of a ChIP- seq peak: Default:200

chipSd

chipSd: A numeric value describing the standard deviation of ChIP-seq peaks. Default: 200

chipSmooth

chipSmooth: A numeric value describing the width of the window used to smooth Occupancy profiles into ChIP profiles. Default:250

stepSize

stepSize: A numeric value describing the step Size (in base pairs) between each ChIP-seq score. Default:10 (Scored every 10 base pairs)

removeBackground

removeBackground: A numeric value describing the value at which score should be removed. Defualt:0 (If negative scores then remove)

noiseFilter

noiseFilter: A character string of one of the following: Zero, Mean, Median, or Sigmoid. Noise filter that will be applied to the ChIP Score during the processingChIP step.

naturalLog

naturalLog: A logical value describing if natural Log will be used to compute the PWM (if FALSE then log2 will be used). Default: TRUE

noOfSites

noOfSites: A Positive integer descibing number of sites (in base pair) should be used from the PFM to compute PWM. Default =0 (Full width of binding site will be used when set to 0)

PWMThreshold

PWMThreshold: Threshold at which PWM Score should be selected (only sites above threshold will be selected - between 0 and 1)

strandRule

strandRule: `mean', `max' or `sum' will dertermine how strand should be handle for computing PWM Scores. Default : `max'

whichstrand

whichstrand: `+',`-' or `+-' on which strand should PWM Score be computed. Default: `+-'

PWMpseudocount

PWMpseudocount: A numeric value describing a PWMpseudocount for PWM computation. Default:1

lambdaPWM

A vector (or single value) contaning values for the ScalingFactorPWM (Also known as lambda).Default:1

Details

ChIPanalyser requires a lot of parameters. parameterOptions was created with the intent of storing and parsing these numerous arguments to the different functions. All parameters in this object are optional although strongly recommend. Some parameters are extracted and updated from function along the pipeline e.g. maxSignal and backgroundSignal are extracted during the processingChIP step. These paramters will be automatically parsed. If you do not which to use them ( or any other parameter) simply parse a new parameterOptions object with your desired paramters.

Value

Returns a parameterOptions with updated values.

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94.

Author

Patrick C. N. Martin <p.martin@essex.ac.uk>

See also

Examples

#
parameterOptions(ploidy = 2, boundMolecules = 1000, backgroundSignal = 0,
    maxSignal = 1, lociWidth = 20000, chipMean = 200, chipSd = 200,
    chipSmooth = 250, stepSize = 10, removeBackground = 0, noiseFilter = "zero",
    naturalLog = TRUE, noOfSites = "all", PWMThreshold = 0.7,
    strandRule = "max", whichstrand = "+-", PWMpseudocount = 1,
    lambdaPWM = 1)
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>                         parameterOptions 
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  processingChIP options 
#>  
#>  chipMean:                                                  200 
#>  chipSd:                                                    200 
#>  chipSmooth:                                                250 
#>  lociWidth:                                               20000 
#>  noiseFilter:                                              zero 
#>  
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  processingChIP options Updated 
#>  
#>  maxSignal:                                                   1 
#>  backgroundSignal:                                            0 
#>  
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  Position Weight Matrix Options 
#>  
#>  naturalLog:                                               TRUE 
#>  noOfSites:                                                 all 
#>  PWMpseudocount:                                              1 
#>  
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  Genome Wide Score options 
#>  
#>  strandRule:                                                max 
#>  whichstrand:                                                +- 
#>  lambdaPWM:                                                   1 
#>  
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  PWM Scores above Threshold options 
#>  
#>  strandRule:                                                max 
#>  whichstrand:                                                +- 
#>  
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  Occupancy options 
#>  
#>  Ploidy:                                                      2 
#>  lambdaPWM:                                                   1 
#>  boundMolecules:                                           1000 
#>  maxSignal:                                                   1 
#>  backgroundSignal:                                            0 
#>  
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
#>  ChIP Profile options 
#>  
#>  chipMean:                                                  200 
#>  chipSd:                                                    200 
#>  chipSmooth:                                                250 
#>  stepSize:                                                   10 
#>  
#> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _