Class "parameterOptions"

parameterOptions is an object used to store and parse the various parameters needed throughout this analysis pipeline.

Objects from the Class

Objects can be created by calls of the form parameterOptions(ploidy, boundMolecules, backgroundSignal, maxSignal, lociWidth, chipMean, chipSd, chipSmooth, stepSize, noiseFilter, removeBackground, lambdaPWM, PWMpseudocount, naturalLog, noOfSites, PWMThreshold, strandRule, whichstrand).

Slots

ploidy:

Object of class "numeric": A numeric Value descibing the ploidy of the organism. Default: 2

boundMolecules:

Object of class "vector": A vector (or single value) containing the number of bound Molecules (bound Transcription Factors): Default: 1000

backgroundSignal:

Object of class "numeric": A numeric value descibing the ChIP-seq background Signal (average signal from real ChIP seq data). Default: 0

maxSignal:

Object of class "numeric": A numeric value describing the highest ChIP-seq signal (from real ChIP-seq data). Default: 1

lociWidth:

Object of class "numeric": A numeric value describing bin size when splitting ChIP seq scores). Default: 20 000

chipMean:

Object of class "numeric": A numeric value describing the mean width of a ChIP- seq peak. Default:150

chipSd:

Object of class "numeric": A numeric value describing the standard deviation of ChIP-seq peaks. Default: 150

chipSmooth:

Object of class "vector": A numeric value describing the width of the window used to smooth Occupancy profiles into ChIP profiles. Default:250

stepSize:

Object of class "numeric": A numeric value describing the step Size (in base pairs) between each ChIP-seq score. Default:10 (Scored every 10 base pairs)

removeBackground:

Object of class "numeric": A numeric value describing the value at which score should be removed. Defualt:0 (If negative scores then remove)

noiseFilter:

Object of class "character" Describes noiseFilter method applied to ChIP scores

PWMThreshold:

Object of class "numeric": Threshold at which PWM Score should be selected (only sites above threshold will be selected - between 0 and 1)

strandRule:

Object of class "character": "mean", "max" or "sum" will dertermine how strand should be handle for computing PWM Scores. Default : "max"

whichstrand:

Object of class "character": "+","-" or "+-" on which strand should PWM Score be computed. Default: "+-"

lambdaPWM:

Object of class "vector" A vector (or single value) contaning values for lambdaPWM Default:1

naturalLog:

Object of class "logical": A logical value describing if natural Log will be used to compute the PWM (if FALSE then log2 will be used). Default: TRUE

noOfSites:

Object of class "nos" A Positive integer descibing number of sites (in base pair) should be used from the PFM to compute PWM. Default =0 (Full width of binding site will be used when set to 0)

PWMpseudocount:

Object of class "numeric": A numeric value describing a PWMpseudocount for PWM computation. Default:1

paramTag:

Object of class "character" ~Internal~

Methods

.paramTag

signature(object = "parameterOptions"): ...

.paramTag<-

signature(object = "parameterOptions", value = "character"): ...

.ZeroBackground

signature(object = "parameterOptions"): ...

.ZeroBackground<-

signature(object = "parameterOptions", value = "vector"): ...

backgroundSignal

signature(object = "parameterOptions"): ...

backgroundSignal<-

signature(object = "parameterOptions", value = "numeric"): ...

boundMolecules

signature(object = "parameterOptions"): ...

boundMolecules<-

signature(object = "parameterOptions", value = "vector"): ...

chipMean

signature(object = "parameterOptions"): ...

chipMean<-

signature(object = "parameterOptions", value = "numeric"): ...

chipSd

signature(object = "parameterOptions"): ...

chipSd<-

signature(object = "parameterOptions", value = "numeric"): ...

chipSmooth

signature(object = "parameterOptions"): ...

chipSmooth<-

signature(object = "parameterOptions", value = "vector"): ...

initialize

signature(.Object = "parameterOptions"): ...

lambdaPWM

signature(object = "parameterOptions"): ...

lambdaPWM<-

signature(object = "parameterOptions", value = "vector"): ...

lociWidth

signature(object = "parameterOptions"): ...

lociWidth<-

signature(object = "parameterOptions", value = "numeric"): ...

maxSignal

signature(object = "parameterOptions"): ...

maxSignal<-

signature(object = "parameterOptions", value = "numeric"): ...

naturalLog

signature(object = "parameterOptions"): ...

naturalLog<-

signature(object = "parameterOptions", value = "logical"): ...

noiseFilter

signature(object = "parameterOptions"): ...

noiseFilter<-

signature(object = "parameterOptions", value = "character"): ...

noOfSites

signature(object = "parameterOptions"): ...

noOfSites<-

signature(object = "parameterOptions", value = "character"): ...

noOfSites<-

signature(object = "parameterOptions", value = "numeric"): ...

ploidy

signature(object = "parameterOptions"): ...

ploidy<-

signature(object = "parameterOptions", value = "numeric"): ...

PWMpseudocount

signature(object = "parameterOptions"): ...

PWMpseudocount<-

signature(object = "parameterOptions", value = "numeric"): ...

PWMThreshold

signature(object = "parameterOptions"): ...

PWMThreshold<-

signature(object = "parameterOptions", value = "numeric"): ...

removeBackground

signature(object = "parameterOptions"): ...

removeBackground<-

signature(object = "parameterOptions", value = "vector"): ...

show

signature(object = "parameterOptions"): ...

stepSize

signature(object = "parameterOptions"): ...

stepSize<-

signature(object = "parameterOptions", value = "numeric"): ...

strandRule

signature(object = "parameterOptions"): ...

strandRule<-

signature(object = "parameterOptions", value = "character"): ...

whichstrand

signature(object = "parameterOptions"): ...

whichstrand<-

signature(object = "parameterOptions", value = "character"): ...

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94. Patrick C.N. Martin and Nicolae Radu Zabe (2020) Dissecting the binding mechanisms of transcription factors to DNA using a statistical thermodynamics framework. CSBJ, 18, 3590-3605.

Author

Partick C. N. Martin <p.martin@essex.ac.uk>

See also

Examples

showClass("parameterOptions")
#> Class "parameterOptions" [package "ChIPanalyser"]
#> 
#> Slots:
#>                                                                           
#> Name:            ploidy   boundMolecules backgroundSignal        maxSignal
#> Class:          numeric           vector          numeric          numeric
#>                                                                           
#> Name:         lociWidth         chipMean           chipSd       chipSmooth
#> Class:          numeric          numeric          numeric           vector
#>                                                                           
#> Name:          stepSize removeBackground      noiseFilter     PWMThreshold
#> Class:          numeric          numeric        character          numeric
#>                                                                           
#> Name:        strandRule      whichstrand        lambdaPWM       naturalLog
#> Class:        character        character           vector          logical
#>                                                          
#> Name:         noOfSites   PWMpseudocount         paramTag
#> Class:              nos          numeric        character
#> 
#> Known Subclasses: "ChIPScore", "genomicProfiles"