Plot Occupancy Profiles

plotOccupancyProfile plots the predicted profiles. If provided, this functions will also plot ChIP-seq profiles, PWMScores (or Occupancy), chromatin States, Goodness of Fit estimates and gene information.

Usage

plotOccupancyProfile(predictedProfile, ChIPScore = NULL,chromatinState = NULL,
    occupancy = NULL,goodnessOfFit = NULL,PWM=FALSE,
    geneRef = NULL,addLegend = TRUE,...)

Arguments

predictedProfile

predictedProfile is a either GRanges containing the predicted profiles for one loci, all loci selected for one paramter, or all loci selected for all parameter combinations selected. (see searchSites)

ChIPScore

ChIPScore is a ChIPscore object containing ChIPscore (or a list of numeric values representing ChIP scores (Experimental ChIP))

chromatinState

chromatinState is a GRanges containing accesible DNA sites or chromatin States.

occupancy

occupancy is a GRanges or a genomicProfiles object contaning PWM scores and Occupancy ( see computeOccupancy)

goodnessOfFit

goodnessOfFit results of the profileAccuracyEstimate function.

PWM

PWM is a logical value that in the case occupancy is provided which of occupancy scores of PWM scores hsould be plotted. Default set at FALSE

geneRef

geneRef is a GRanges containing gene information on exons,introns, UTR's, enhancers or any other genetic element to be plotted.

addLegend

addLegend is a logical value defining if the legend should be added. The legend will add all elements provided. See details.

...

Any other graphical Parameter of the following : cex, cex.lab, cex.main, densityCS , densityGR , ylab, xlab, main, colPred, colChIP, colOccup, colCS, colGR, n_axis_ticks. See details.

Details

Once the predicted ChIP-seq like profiles have been computed, it is possible to plot these profiles.

This functions allows to control graphical parameters. In short:

* col = color values - exact number of colors or colors that will be used in a colorRampPalettte.

* cex = font sizes - for text, axis labels and main

* Density = fill density for chromatin state and/or geneRef blocks

Pred = predictedProfile ChIP = ChIP score (Experimental ChIP data) CS = Chromatin States GR = Gene reference Occup = Occupnacy locations

Value

Returns a profile plot with "Occupancy" on the y axis and DNA position on the the X- axis.

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94. Patrick C.N. Martin and Nicolae Radu Zabe (2020) Dissecting the binding mechanisms of transcription factors to DNA using a statistical thermodynamics framework. CSBJ, 18, 3590-3605.

Author

Patrick C.N. Martin <pcnmartin@gmail.com>

Examples

#Data extraction
data(ChIPanalyserData)
# path to Position Frequency Matrix
PFM <- file.path(system.file("extdata",package="ChIPanalyser"),"BEAF-32.pfm")
#As an example of genome, this example will run on the Drosophila genome
# \donttest{
if(!require("BSgenome.Dmelanogaster.UCSC.dm6", character.only = TRUE)){
    if (!requireNamespace("BiocManager", quietly=TRUE))
        install.packages("BiocManager")
    BiocManager::install("BSgenome.Dmelanogaster.UCSC.dm6")
    }
library(BSgenome.Dmelanogaster.UCSC.dm6)
DNASequenceSet <- getSeq(BSgenome.Dmelanogaster.UCSC.dm6)
#Building data objects
GPP <- genomicProfiles(PFM=PFM,PFMFormat="JASPAR", BPFrequency=DNASequenceSet)

# Computing Genome Wide
GenomeWide <- computeGenomeWideScores(DNASequenceSet = DNASequenceSet,
    genomicProfiles = GPP)
#> Extracting genome wide scores 
#> Whole Genome ~ Both strands 
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Warning: 'subject' contains letters not in [ACGT] ==> assigned weight 0 to them
#> Computing Mean waiting time 

#Compute PWM Scores
PWMScores <- computePWMScore(DNASequenceSet = DNASequenceSet,
    genomicProfiles = GenomeWide,
    loci = top, chromatinState = Access)
#> PWM Scores Extraction
#Compute Occupnacy
Occupancy <- computeOccupancy(genomicProfiles = PWMScores)
#> Computing Occupancy at sites higher than threshold. 

#Compute ChIP profiles
chipProfile <- computeChipProfile(loci = top,
    genomicProfiles = Occupancy)
#> Error in computeChipProfile(loci = top, genomicProfiles = Occupancy): could not find function "computeChipProfile"

#Plotting Profile
plotOccupancyProfile(predictedProfile=chipProfile,
    ChIPScore = chip,
    chromatinState = Access,
    occupancy = Occupancy,
    geneRef =geneRef)
#> Error in .is.genomicProfiles(predictedProfile): object 'chipProfile' not found

plotOccupancyProfile(predictedProfile=chipProfile,
    ChIPScore = chip,
    chromatinState = Access,
    occupancy = Occupancy,
    geneRef = geneRef,
    colCS = c("red","blue"),
    densityGR = 60)
#> Error in .is.genomicProfiles(predictedProfile): object 'chipProfile' not found

# }