Estimating Accuracy of predicted Profiles

profileAccuracyEstimate will compare the predicted ChIP-seq-like profile to real ChIP-seq data and return a set of metrics describing how accurate the predicted model is compared to real data.

Usage

profileAccuracyEstimate(genomicProfiles,ChIPScore,
      parameterOptions=NULL,method="all",cores=1)

Arguments

genomicProfiles

genomicProfiles is the result of computeChIPProfile

ChIPScore

ChIPScore is the result of processingChIP. Extracted/Normalised experimental ChIP scores.

parameterOptions

parameterOptions is a parameterOptions object for paramter specification.

method

method is the method that will be used to assess model quality agianst ChIP-seq data. Method can be one of the following: pearson, spearman, kendall, ks, geometric,fscore, MSE,or all.Fscore contains f-score, precision,recall, MCC, Accuracy and AUC ROC.

cores

cores is the number of cores used to extract ChIP scores. Default = 1

Details

In order to assess the quality of the model against experimental ChIP-seq data, ChIPanalyser offers a wide range of method to choose from. These methods are also used when computing optimal paramters.

Value

Returns list of goodness of fit metrics for each loci and each parameter selected.

References

Zabet NR, Adryan B (2015) Estimating binding properties of transcription factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94. Patrick C.N. Martin and Nicolae Radu Zabe (2020) Dissecting the binding mechanisms of transcription factors to DNA using a statistical thermodynamics framework. CSBJ, 18, 3590-3605.

Author

Patrick C. N. Martin <pm16057@essex.ac.uk>

Examples

#Data extraction
data(ChIPanalyserData)
# path to Position Frequency Matrix
PFM <- file.path(system.file("extdata",package="ChIPanalyser"),"BEAF-32.pfm")
#As an example of genome, this example will run on the Drosophila genome
# \donttest{
  if(!require("BSgenome.Dmelanogaster.UCSC.dm6", character.only = TRUE)){
      if (!requireNamespace("BiocManager", quietly=TRUE))
          install.packages("BiocManager")
      BiocManager::install("BSgenome.Dmelanogaster.UCSC.dm6")
      }
  library(BSgenome.Dmelanogaster.UCSC.dm6)
  DNASequenceSet <- getSeq(BSgenome.Dmelanogaster.UCSC.dm6)
  # Building genomicProfiles object
  GPP <- genomicProfiles(PFM=PFM,PFMFormat="JASPAR", BPFrequency=DNASequenceSet)

  # Computing Genome Wide
  GenomeWide <- computeGenomeWideScore(genomicProfiles = GPP,
      DNASequenceSet = DNASequenceSet)
#> Error in computeGenomeWideScore(genomicProfiles = GPP, DNASequenceSet = DNASequenceSet): could not find function "computeGenomeWideScore"

  #Compute PWM Scores
  PWMScores <- computePWMScore(genomicProfiles = GenomeWide,
       DNASequenceSet = DNASequenceSet, loci = top, chromatinState = Access)
#> Error in .is.genomicProfiles(genomicProfiles): object 'GenomeWide' not found
  #Compute Occupnacy
  Occupancy <- computeOccupancy(genomicProfiles = PWMScores)
#> Error in .is.genomicProfiles(genomicProfiles): object 'PWMScores' not found

  #Compute ChIP profiles
  chipProfile <- computeChIPProfile(genomicProfiles=Occupancy,loci=top)
#> Error in .is.genomicProfiles(genomicProfiles): object 'Occupancy' not found
#Estimating accuracy estimate
AccuracyEstimate <- profileAccuracyEstimate(genomicProfiles = chipProfile,
    ChIPScore = chip,
    occupancyProfileParameters = OPP)
#> Error in profileAccuracyEstimate(genomicProfiles = chipProfile, ChIPScore = chip,     occupancyProfileParameters = OPP): unused argument (occupancyProfileParameters = OPP)
# }